main-node/pyproject.toml
[project]
name = "main-node"
version = "0.1.0"
description = "RTX 5070 Ti Computer Vision Node"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "torchvision",
    "opencv-python",
    "ultralytics",
    "PySide6>=6.8.0",
    "eclipse-zenoh>=1.8.0",
    "python-dotenv>=1.2.2",
]

[tool.uv]
environments = ["sys_platform == 'win32'"]
index-strategy = "unsafe-first-match"

[[tool.uv.index]]
name = "pytorch-nightly"
url = "https://download.pytorch.org/whl/nightly/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-nightly" }
torchvision = { index = "pytorch-nightly" }